fix(settings): update suggestions settings
