+++ src/utils/i18n/locales/en/settings.json
@@ -40,16 +40,8 @@
 		"speed_fee_custom": "Set Custom Fee",
 		"speed_fee_total": "{feeSats} sats for the average transaction",
 		"speed_fee_total_fiat": "{feeSats} sats for the average transaction ({fiatSymbol}{fiatFormatted})",
-		"suggestions": "Suggestions",
-		"suggestions_visible": "Visible",
-		"suggestions_hidden": "Hidden",
-		"suggestions_display": "Display suggestions",
-		"suggestions_reset": "Reset Suggestions",
 		"tags": "Tags",
 		"tags_previously": "Previously used tags",
-		"reset_title": "Reset Suggestions?",
-		"reset_desc": "Are you sure you want to reset the suggestions? They will reappear in case you have removed them from your Bitkit wallet overview.",
-		"reset_confirm": "Yes, Reset",
 		"widgets": "Widgets"
 	},
 	"security": {
@@ -107,6 +99,10 @@
 		"pp_contacts_switch": "Enable payments with contacts*",
 		"address_viewer": "Address Viewer",
 		"rescan": "Rescan Addresses",
+		"suggestions_reset": "Reset Suggestions",
+		"reset_title": "Reset Suggestions?",
+		"reset_desc": "Are you sure you want to reset the suggestions? They will reappear in case you have removed them from your Bitkit wallet overview.",
+		"reset_confirm": "Yes, Reset",
 		"lightning_connections": "Lightning Connections",
 		"lightning_node": "Lightning Node",
 		"electrum_server": "Electrum Server",
